minor bug fixes. tmp disable graphviz
git-svn-id: https://oceanai.mit.edu/svn/moos-ivp-aro/trunk@7038 7b6de600-d9ab-45c9-a1ca-e8145d3c6daf

+++ ivp/src/app_alogview/GUI_VarScope.cpp
@@ -363,7 +363,7 @@ void GUI_VarScope::cb_ButtonSetVar(Fl_Widget* o, int v) {
 inline void GUI_VarScope::cb_ButtonDelVar_i(int mix) {
   vector<unsigned int> new_current_mix_vars;
   for(unsigned int i=0; i<m_current_mix_vars.size(); i++) {
-    if(m_current_mix_vars[i] == mix)
+    if(m_current_mix_vars[i] == (unsigned int)(mix))
       m_vsmodel.delVarPlot(mix);
     else
       new_current_mix_vars.push_back(m_current_mix_vars[i]);
